// crates/aegis-web/frontend/src/types/Task.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskCreator } from "./TaskCreator";
import type { TaskStatus } from "./TaskStatus";

export type Task = { task_id: string, description: string, status: TaskStatus, assigned_agent_id: string | null, created_by: TaskCreator, created_at: string, completed_at: string | null, receipt_path: string | null, };